WIP: generate oneof types
For each oneof group in a message,
we create a proxy type that sets one field in the group.
These types will be later be used by toolkit to flatten
oneof fields.

Testing with pubsub, the generated code compiles and looks
good enough for early feedback.
However, we should test with other APIs, and there are
some finer points I'd like to discuss more, hence WIP.

The oneof types are nested in the message type,
because the oneof group is scoped in the message in protobuf.