Support deprecated resource names (#64)
Support deprecated_collections in gapic config v2. If the pattern of a deprecated_collection belong to a multi-pattern resource in proto files, generate a resource name class based on the entity_name and the pattern and mark the generated class as deprecated. This way we won't break existing clients with multi-pattern resource names.

Stop generating entity names from patterns by constructing a trie. We do not need it in the new generated multi-pattern resource name.

Stop guessing the parent resource of a resource referenced by child_type. We only generate the resource if there is a parent resource explicitly defined in the protos, with each of whose pattern(s) is a parent of each of the pattern(s) of the resource referenced by child_type. This is consistent with C# micro generator, and with the file-level annotations we shouldn't need to guess the parent resource any more.

Remove google-gax as a dependency since it does not seem to be used any more.