fix: stop generating `parseList` and `toStringList` if a multi-pattern resource name has subclasses (#79)
In Java, parent classes and child classes cannot have methods with the same signature if the methods:

both have generic types in parameters or return type
the methods have the same type eraser.
The rule affects libraries migrated from gapic config. We used to generate super-class and subclass for multi-pattern resource names for these libraries and we need to continue to do so for backward-compatibility. Therefore we do not generate toStringList and parseList on the surface of the new multi-pattern resource names if a library is converted from gapic config.

We check if a resource is carried over from gapic v1 by checking whether there is a deprecated_collection that matches one of its patterns in gapic v2.